Moved app resources into xaml
