--- Todo/Todo/App.cs
@@ -1,52 +1,0 @@
-﻿using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-[assembly: XamlCompilation(XamlCompilationOptions.Compile)]
-namespace Todo
-{
-    public class App : Application
-	{
-		static TodoItemDatabase database;
-
-		public App()
-		{
-			Resources = new ResourceDictionary();
-			Resources.Add("primaryGreen", Color.FromHex("91CA47"));
-			Resources.Add("primaryDarkGreen", Color.FromHex("6FA22E"));
-
-			var nav = new NavigationPage(new TodoListPage());
-			nav.BarBackgroundColor = (Color)App.Current.Resources["primaryGreen"];
-			nav.BarTextColor = Color.White;
-
-			MainPage = nav;
-		}
-
-		public static TodoItemDatabase Database
-		{
-			get
-			{
-				if (database == null)
-				{
-                    database = new TodoItemDatabase();
-				}
-				return database;
-			}
-		}
-
-		protected override void OnStart()
-		{
-
-		}
-
-		protected override void OnSleep()
-		{
-
-		}
-
-		protected override void OnResume()
-		{
-
-		}
-	}
-}
-
