+++ Todo/Todo/App.xaml
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Application xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="Todo.App">
+    <Application.Resources>
+        <ResourceDictionary>
+            <Color x:Key="primaryGreen">#91CA47</Color>
+            <Color x:Key="primaryDarkGreen">#6FA22E</Color>
+        </ResourceDictionary>
+    </Application.Resources>
+</Application>
