+++ Todo/Todo/Views/TodoItemPage.xaml.cs
@@ -3,30 +3,30 @@ using Xamarin.Forms;
 
 namespace Todo
 {
-	public partial class TodoItemPage : ContentPage
-	{
-		public TodoItemPage()
-		{
-			InitializeComponent();
-		}
+    public partial class TodoItemPage : ContentPage
+    {
+        public TodoItemPage()
+        {
+            InitializeComponent();
+        }
 
-		async void OnSaveClicked(object sender, EventArgs e)
-		{
-			var todoItem = (TodoItem)BindingContext;
-			await App.Database.SaveItemAsync(todoItem);
-			await Navigation.PopAsync();
-		}
+        async void OnSaveClicked(object sender, EventArgs e)
+        {
+            var todoItem = (TodoItem)BindingContext;
+            await App.Database.SaveItemAsync(todoItem);
+            await Navigation.PopAsync();
+        }
 
-		async void OnDeleteClicked(object sender, EventArgs e)
-		{
-			var todoItem = (TodoItem)BindingContext;
-			await App.Database.DeleteItemAsync(todoItem);
-			await Navigation.PopAsync();
-		}
+        async void OnDeleteClicked(object sender, EventArgs e)
+        {
+            var todoItem = (TodoItem)BindingContext;
+            await App.Database.DeleteItemAsync(todoItem);
+            await Navigation.PopAsync();
+        }
 
-		async void OnCancelClicked(object sender, EventArgs e)
-		{
-			await Navigation.PopAsync();
-		}
-	}
+        async void OnCancelClicked(object sender, EventArgs e)
+        {
+            await Navigation.PopAsync();
+        }
+    }
 }
